+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Atualizações
 
+## [Versão 3.49.82]
+- Relatório de CPF, RG e NIS do Alunos por Turma
+- Relatório de CPF, RG e NIS do Alunos por Escola
+- Relatório de CPF, RG e NIS do Alunos de Todas as Escolas
+
 ## [Versão 3.46.82]
 - Relatório de Alunos Benefeciários do Bolsa Família por Turma
 - Relatório de Alunos Benefeciários do Bolsa Família de Todas as Turmas